Refactor: Rename fs.ts -> filesystem.ts
